requirements.txt
notebook
torch

numpy
matplotlib
seaborn

pillow
imageio
tifffile
scipy

scikit-image
scikit-learn

opencv-python


sphinx

flask